@@ -2,6 +2,7 @@
   "name": "ajs",
   "version": "1.0.0",
   "description": "",
+  "type":"commonjs",
   "main": "index.js",
   "scripts": {
     "start": "webpack serve --mode development",
@@ -27,7 +28,7 @@
     "@typescript-eslint/parser": "^6.13.2",
     "babel-loader": "^9.1.3",
     "css-loader": "^6.8.1",
-    "eslint": "^8.55.0",
+    "eslint": "^9.34.0",
     "eslint-plugin-jest": "^27.4.2",
     "html-loader": "^4.2.0",
     "html-webpack-plugin": "^5.5.3",
